repo_path: move RepoPathUiConverter to separate module
The `RepoPathUiConverter` type should perhaps not be moved to the
coming `jj-core` crate, so this patch prepares for moving `RepoPath`
to the crate without leaving a `repo_path` module without `RepoPath`
behind.
